+++ CHANGELOG.md
@@ -9,6 +9,19 @@ Unreleased section should follow [Release Toolkit](https://github.com/newrelic/r
 
 ## Unreleased
 
+### enhancements
+- Introduced Query Performance Monitoring
+- Enabled reporting for Slow Running Queries
+- Added detailed Query Execution Plan analysis for Slow Running Queries
+- Added Reporting for Wait Events
+- Added Reporting for Blocking Sessions
+- Add FIPS compliant packages
+
+## v2.16.1 - 2025-01-20
+
+### ⛓️ Dependencies
+- Updated golang patch version to v1.23.5
+
 ## v2.16.0 - 2024-11-11
 
 ### 🚀 Enhancements